fixed max score in leaderboard

+++ models/player.go
@@ -5,5 +5,5 @@ import "go.mongodb.org/mongo-driver/bson/primitive"
 type Player struct {
 	Id     primitive.ObjectID `json:"id" bson:"_id"`
 	Name   string             `json:"name,omitempty" validate:"required"`
-	Scores []int              `json:"scores"`
+	Scores []float32          `json:"scores"`
 }
